Move api/apitest to top level
