+++ sdk/trace/trace_test.go
@@ -27,13 +27,13 @@ import (
 
 	"go.opentelemetry.io/otel/api/global"
 	"go.opentelemetry.io/otel/label"
+	"go.opentelemetry.io/otel/oteltest"
 
 	"github.com/google/go-cmp/cmp"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	grpccodes "google.golang.org/grpc/codes"
 
-	"go.opentelemetry.io/otel/api/apitest"
 	"go.opentelemetry.io/otel/api/trace"
 	apitrace "go.opentelemetry.io/otel/api/trace"
 	otelcodes "go.opentelemetry.io/otel/codes"
@@ -61,7 +61,7 @@ func init() {
 
 func TestTracerFollowsExpectedAPIBehaviour(t *testing.T) {
 	tp := NewProvider(WithConfig(Config{DefaultSampler: TraceIDRatioBased(0)}))
-	harness := apitest.NewHarness(t)
+	harness := oteltest.NewHarness(t)
 	subjectFactory := func() trace.Tracer {
 		return tp.Tracer("")
 	}
